Clear commit with navigation
