+++ app/src/main/java/bav/anyway/MainViewModel.kt
@@ -1,0 +1,9 @@
+package bav.anyway
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import javax.inject.Inject
+
+class MainViewModel
+@Inject constructor(app: Application) : AndroidViewModel(app) {
+}
