+++ app/src/main/java/bav/anyway/di/ScreenBindingModule.kt
@@ -1,6 +1,8 @@
 package bav.anyway.di
 
+import bav.anyway.FirstScreenFragment
 import bav.anyway.MainActivity
+import bav.anyway.MapFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -8,4 +10,10 @@ import dagger.android.ContributesAndroidInjector
 abstract class ScreenBindingModule {
     @ContributesAndroidInjector
     abstract fun bindMainActivity(): MainActivity
+
+    @ContributesAndroidInjector
+    abstract fun bindFirstScreenFragment(): FirstScreenFragment
+
+    @ContributesAndroidInjector
+    abstract fun bindMapFragment(): MapFragment
 }
