+++ app/src/main/res/layout/activity_main.xml
@@ -14,4 +14,5 @@
             android:layout_height="match_parent"
             app:defaultNavHost="true"
             app:navGraph="@navigation/main_graph" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
