+++ app/src/main/res/layout/fragment_first_screen.xml
@@ -11,10 +11,9 @@
 
     <TextView
             android:id="@+id/RequestTextView"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
             android:layout_marginTop="150dp"
-            android:layout_marginBottom="114dp"
             android:text="Введите запрос"
             android:textAlignment="center"
             android:textSize="20dp"
@@ -28,10 +27,9 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="150dp"
-            android:layout_marginBottom="383dp"
+            android:layout_marginTop="100dp"
             android:clickable="true"
             android:text="Go"
-            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/RequestTextView" />
